+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 // app/page.tsx
 import React from 'react'
-import AppContent from '../appContent'
+import AppContent from '../components/appContent'
 
 export default function Home() {
   return <AppContent />
