+++ src/hooks/useElementDragEffect.tsx
@@ -16,7 +16,7 @@
  */
 
 import { useState, useEffect, useCallback, useRef } from 'react';
-import { Element } from '../types';
+import { Element } from '../types/types';
 import { useCanvas } from '../context/canvasContext';
 import { isDescendant } from '../utils/elementHelpers';
 import { ToastMessages } from '../constants/toastMessages';
