+++ src/state/__test__/dragAndDrop.test.ts
@@ -1,7 +1,7 @@
 // src/state/__test__/dragAndDrop.test.ts
 import { renderHook, act } from '@testing-library/react';
 import { useStore } from './textUtils';
-import { Element } from '../../types';
+import { Element } from '../../types/types';
 
 
 describe('ドラッグ＆ドロップ', () => {
